+++ git/refs/symbolic.py
@@ -308,7 +308,7 @@ class SymbolicReference(object):
 
         lfd = LockedFD(fpath)
         fd = lfd.open(write=True, stream=True)
-        fd.write(write_value)
+        fd.write(write_value.encode('ascii'))
         lfd.commit()
 
         # Adjust the reflog